// senegal-example-app/angular-frontend/senegal-frontend/src/app/author/api/search-author-instruction-to.model.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {AuthorIdTO} from "./author-id-to.model";


export interface SearchAuthorInstructionTO {
    authorId: AuthorIdTO | undefined;
    firstname: string | undefined;
    lastname: string | undefined;
}